wip: cleaning up code and start refactoring before release
